Inject downstream accumulator.
